refactor: 뷰 파일 생성
