+++ AND-SOPT-iOS/AND-SOPT-iOS/Assignment/Week02/View/AppInfoSummaryView.swift
@@ -1,0 +1,175 @@
+//
+//  AppInfoSummaryView.swift
+//  AND-SOPT-iOS
+//
+//  Created by 김예지 on 10/24/24.
+//
+
+import UIKit
+
+class AppInfoSummaryView: UIView {
+    // MARK: 평가 개수, 앱 수상 내역, 연령 정보
+    
+    private let infoStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .horizontal
+        stackView.distribution = .fillProportionally
+        //        stackView.distribution = .fillEqually
+        stackView.alignment = .fill
+        return stackView
+    }()
+    
+    // 평가
+    private let rateStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.spacing = 5
+        stackView.alignment = .center
+        return stackView
+    }()
+    
+    private let rateCountLabel: UILabel = {
+        let label = UILabel()
+        label.text = "8.4만개의 평가"
+        label.font = .systemFont(ofSize: 13, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let rateTextLabel: UILabel = {
+        let label = UILabel()
+        label.text = "4.4"
+        label.font = .systemFont(ofSize: 18, weight: .black)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let rateStarLabel: UILabel = {
+        let label = UILabel()
+        label.text = "★★★★☆"
+        label.font = .systemFont(ofSize: 15, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let verticalLine1: UIView = {
+        let view = UIView()
+        view.backgroundColor = .lightGray
+        return view
+    }()
+    
+    
+    // 수상
+    private let awardStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.spacing = 5
+        stackView.alignment = .center
+        return stackView
+    }()
+    
+    private let awardLabel: UILabel = {
+        let label = UILabel()
+        label.text = "수상"
+        label.font = .systemFont(ofSize: 13, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let awardTitleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "에디터의\n선택"
+        label.numberOfLines = 2
+        label.textAlignment = .center
+        label.font = .systemFont(ofSize: 12, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let awardCategoryLabel: UILabel = {
+        let label = UILabel()
+        label.text = "앱"
+        label.font = .systemFont(ofSize: 15, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let verticalLine2: UIView = {
+        let view = UIView()
+        view.backgroundColor = .lightGray
+        return view
+    }()
+    
+    // 연령
+    private let ageStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.spacing = 5
+        stackView.alignment = .center
+        return stackView
+    }()
+    
+    private let ageTitleLabel: UILabel = {
+        let label = UILabel()
+        label.text = "연령"
+        label.font = .systemFont(ofSize: 15, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let ageTextLabel: UILabel = {
+        let label = UILabel()
+        label.text = "4+"
+        label.font = .systemFont(ofSize: 20, weight: .black)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    private let ageUnitLabel: UILabel = {
+        let label = UILabel()
+        label.text = "세"
+        label.font = .systemFont(ofSize: 15, weight: .bold)
+        label.textColor = .systemGray
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setStyle()
+        setUI()
+        setLayout()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setStyle(){
+        self.backgroundColor = .black
+    }
+    
+    private func setUI() {
+        rateStackView.addArrangedSubviews(rateCountLabel, rateTextLabel, rateStarLabel)
+        awardStackView.addArrangedSubviews(awardLabel, awardTitleLabel, awardCategoryLabel)
+        ageStackView.addArrangedSubviews(ageTitleLabel, ageTextLabel, ageUnitLabel)
+        infoStackView.addArrangedSubviews(rateStackView, verticalLine1, awardStackView, verticalLine2, ageStackView)
+        
+        addSubviews(infoStackView)
+    }
+    
+    private func setLayout() {
+        infoStackView.snp.makeConstraints {
+            $0.top.equalToSuperview()
+            $0.leading.trailing.equalToSuperview()
+        }
+        
+        verticalLine1.snp.makeConstraints {
+            $0.width.equalTo(0.3)
+        }
+        
+        verticalLine2.snp.makeConstraints {
+            $0.width.equalTo(0.3)
+        }
+    }
+    
+}
